Resolved merge conflicts Updated dependencies Refactored Task files Reset prompt modal after Task creation
