+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@react-native-community/viewpager": "^5.0.11",
+    "@react-navigation/drawer": "^6.6.6",
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/native-stack": "^6.9.17",
     "@react-navigation/stack": "^6.3.20",
@@ -17,12 +18,13 @@
     "@rneui/themed": "^4.0.0-rc.8",
     "expo": "~49.0.15",
     "expo-status-bar": "~1.6.0",
+    "firebase": "^10.6.0",
     "react": "18.2.0",
     "react-native": "0.72.6",
     "react-native-gesture-handler": "~2.12.0",
     "react-native-pager-view": "6.2.0",
     "react-native-reanimated": "~3.3.0",
-    "react-native-safe-area-context": "^4.6.3",
+    "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0"
   },
   "devDependencies": {
