+++ src/screen/CreateTask.js
@@ -1,64 +1,7 @@
-//import React, { useState } from 'react';
-//import { View, Text, StyleSheet, TouchableOpacity, TextInput, Button, Modal } from 'react-native';
-//import PromptTask from './TaskInfo';
+import React, { useState } from 'react';
+import { View, Text, StyleSheet, TouchableOpacity, TextInput, Button, Modal, ScrollView } from 'react-native';
 
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-const CreateTask = (props) => {
-    return (
-        <View style = {styles.item}>
-            <Text style={styles.itemText}>{props.text}</Text>
-        </View>
-    )
-}
-
-const styles = StyleSheet.create({
-    item: {
-        backgroundColor: '#dc143c',
-        padding: 20,
-        alignItems: 'center',
-        marginTop: 50,
-        borderRadius: 15,
-    },
-    itemText: {
-        fontWeight: 'bold',
-    }
-})
-
-export default CreateTask;
-
-
-/* const Task = () => {
-  const [tasks, setTask] = useState([]);
-
-  const addTask = (title, description, notes) => {
-    const counter = tasks.length;
-    setTask([...tasks, {title, description, notes}]);
-  };
-  return (
-    <View style={styles.container}>
-      <View style={styles.tasksWrapper}>
-        <PromptTask addTask={addTask} />
-        <Text style = {styles.sectionTitle}>Tasks</Text>
-        <ScrollView>
-          {tasks.map((task, key) => (
-            <CreateTask 
-              key={key} 
-              title={task.title}
-              description={task.description}
-              notes={task.notes}
-            />
-          ))}
-        </ScrollView>
-      </View>
-      <MainFile />
-    </View>
-  );
-}
-
-/* const CreateTask = (props) => {
-    //const [properties, setProps] = useState();
+export default CreateTask = (props) => {
     const [modalVisible, setModalVisible] = useState(false);
 
     const title = props.title;
@@ -93,8 +36,8 @@ export default CreateTask;
             </TouchableOpacity>
         </View>
     )
-} */
-/*
+}
+
 const styles = StyleSheet.create({
     centeredView: {
         flex: 1,
@@ -153,6 +96,3 @@ const styles = StyleSheet.create({
         fontWeight: 'bold',
     },
 })
-
-export default Task;
-*/
